Move ffi.rs into risc0-sys remove explicit linking
