+++ risc0/circuit/rv32im/src/lib.rs
@@ -21,8 +21,6 @@ mod cpp;
 pub mod cpu;
 #[cfg(feature = "cuda")]
 pub mod cuda;
-#[cfg(feature = "prove")]
-mod ffi;
 mod info;
 #[cfg(feature = "metal")]
 pub mod metal;
