+++ risc0/r0vm/Cargo.toml
@@ -12,7 +12,7 @@ anyhow = "1.0"
 bytemuck = "1.12"
 clap = { version = "4.0", features = ["derive"] }
 env_logger = "0.10.0"
-risc0-zkvm = { workspace = true, features = ["default", "profiler"] }
+risc0-zkvm = { workspace = true, features = ["default", "profiler", "binfmt"] }
 
 [dev-dependencies]
 anyhow = "1.0"
