+++ risc0/sys/Cargo.toml
@@ -8,6 +8,9 @@ homepage = { workspace = true }
 repository = { workspace = true }
 links = "risc0-sys"
 
+[dependencies]
+risc0-core = { workspace = true }
+
 [build-dependencies]
 cc = { version = "1.0", features = ["parallel"] }
 glob = "0.3"
